Дополнен UI и прочии мелкие улучшения

+++ src/AppStats.h
@@ -7,7 +7,7 @@ struct AppStats {
     QString exe_filename;
     QString display_name;
     bool is_hidden;
-    int64_t category_id;
+    int category_id;
     int64_t total_time;
 };
